# Use double for maximum and minimum zoom level (#752)
* # Use double for maximum and minimum zoom level

* #734 Change int to double in wrapper mapview

+++ osmdroid-android/src/main/java/org/osmdroid/api/IMapView.java
@@ -18,7 +18,7 @@ public interface IMapView {
 	 * @since 6.0
 	 */
 	double getZoomLevelDouble();
-	int getMaxZoomLevel();
+	double getMaxZoomLevel();
      @Deprecated
 	int getLatitudeSpan();
      @Deprecated
