+++ osmdroid-android/src/main/java/org/osmdroid/views/MapView.java
@@ -97,8 +97,8 @@ public class MapView extends ViewGroup implements IMapView, MapViewConstants,
 	protected final AtomicReference<Double> mTargetZoomLevel = new AtomicReference<>();
 	protected final AtomicBoolean mIsAnimating = new AtomicBoolean(false);
 
-	protected Integer mMinimumZoomLevel;
-	protected Integer mMaximumZoomLevel;
+	protected Double mMinimumZoomLevel;
+	protected Double mMaximumZoomLevel;
 
 	private final MapController mController;
 
@@ -394,8 +394,8 @@ public class MapView extends ViewGroup implements IMapView, MapViewConstants,
 	 * Used to be an int - is a double since 6.0
 	 */
 	double setZoomLevel(final double aZoomLevel) {
-		final int minZoomLevel = getMinZoomLevel();
-		final int maxZoomLevel = getMaxZoomLevel();
+		final double minZoomLevel = getMinZoomLevel();
+		final double maxZoomLevel = getMaxZoomLevel();
 
 		final double newZoomLevel = Math.max(minZoomLevel, Math.min(maxZoomLevel, aZoomLevel));
 		final double curZoomLevel = this.mZoomLevel;
@@ -515,7 +515,7 @@ public class MapView extends ViewGroup implements IMapView, MapViewConstants,
 	/**
 	 * Get the minimum allowed zoom level for the maps.
 	 */
-	public int getMinZoomLevel() {
+	public double getMinZoomLevel() {
 		return mMinimumZoomLevel == null ? mMapOverlay.getMinimumZoomLevel() : mMinimumZoomLevel;
 	}
 
@@ -523,7 +523,7 @@ public class MapView extends ViewGroup implements IMapView, MapViewConstants,
 	 * Get the maximum allowed zoom level for the maps.
 	 */
 	@Override
-	public int getMaxZoomLevel() {
+	public double getMaxZoomLevel() {
 		return mMaximumZoomLevel == null ? mMapOverlay.getMaximumZoomLevel() : mMaximumZoomLevel;
 	}
 
@@ -531,7 +531,7 @@ public class MapView extends ViewGroup implements IMapView, MapViewConstants,
 	 * Set the minimum allowed zoom level, or pass null to use the minimum zoom level from the tile
 	 * provider.
 	 */
-	public void setMinZoomLevel(Integer zoomLevel) {
+	public void setMinZoomLevel(Double zoomLevel) {
 		mMinimumZoomLevel = zoomLevel;
 	}
 
@@ -539,12 +539,12 @@ public class MapView extends ViewGroup implements IMapView, MapViewConstants,
 	 * Set the maximum allowed zoom level, or pass null to use the maximum zoom level from the tile
 	 * provider.
 	 */
-	public void setMaxZoomLevel(Integer zoomLevel) {
+	public void setMaxZoomLevel(Double zoomLevel) {
 		mMaximumZoomLevel = zoomLevel;
 	}
 
 	public boolean canZoomIn() {
-		final int maxZoomLevel = getMaxZoomLevel();
+		final double maxZoomLevel = getMaxZoomLevel();
 		if ((isAnimating() ? mTargetZoomLevel.get() : mZoomLevel) >= maxZoomLevel) {
 			return false;
 		}
@@ -552,7 +552,7 @@ public class MapView extends ViewGroup implements IMapView, MapViewConstants,
 	}
 
 	public boolean canZoomOut() {
-		final int minZoomLevel = getMinZoomLevel();
+		final double minZoomLevel = getMinZoomLevel();
 		if ((isAnimating() ? mTargetZoomLevel.get() : mZoomLevel) <= minZoomLevel) {
 			return false;
 		}
