+++ osmdroid-third-party/src/main/java/org/osmdroid/google/wrapper/MapView.java
@@ -75,7 +75,7 @@ public class MapView implements IMapView {
 	}
 
 	@Override
-	public int getMaxZoomLevel() {
+	public double getMaxZoomLevel() {
 		return mMapView.getMaxZoomLevel();
 	}
